feat: 반응형 옵션 추가
