+++ src/lib/InfiniteSlide/InfiniteSlide.tsx
@@ -1,7 +1,6 @@
-import React, { useEffect, useLayoutEffect, useRef, useState } from 'react';
+import React, { useLayoutEffect, useState } from 'react';
+import InnerInfiniteSlide from '../InnerInfiniteSlide/InnerInfiniteSlide';
 import { IProps } from './InfiniteSlide.types';
-import GlobalStyle from '../../styles/GlobalStyle';
-import * as S from './InfiniteSlide.styles';
 
 const InfiniteSlide: React.FC<IProps> = ({
   slidesToScroll = 1,
@@ -10,94 +9,32 @@ const InfiniteSlide: React.FC<IProps> = ({
   rightArrow,
   auto = false,
   interval = 4,
+  responsive = [],
 }) => {
-  const [currentSlide, setCurrentSlide] = useState(1);
-  const [isTransitioning, setIsTransitioning] = useState(false);
-  const [slideWidth, setSlideWidth] = useState(0);
-
-  const slideRef = useRef<HTMLUListElement>(null);
-  const containerRef = useRef<HTMLDivElement>(null);
-  const timerRef = useRef<NodeJS.Timeout>();
-
-  const slides = React.Children.toArray(children);
-  const DATA = [...slides.slice(-slidesToScroll), ...slides, ...slides.slice(0, slidesToScroll)];
-
-  const handleNextSlide = () => {
-    if (isTransitioning) return;
-    setIsTransitioning(true);
-    if (currentSlide !== slides.length && currentSlide + 2 * slidesToScroll > slides.length + slidesToScroll) {
-      setCurrentSlide(slides.length);
-    } else {
-      setCurrentSlide((prev) => prev + slidesToScroll);
-    }
-  };
-
-  const handlePrevSlide = () => {
-    if (isTransitioning) return;
-    setIsTransitioning(true);
-    if (currentSlide !== slidesToScroll && currentSlide - slidesToScroll < slidesToScroll) {
-      setCurrentSlide(slidesToScroll);
-    } else {
-      setCurrentSlide((prev) => prev - slidesToScroll);
-    }
-  };
-
-  const handleTransitionEnd = () => {
-    setIsTransitioning(false);
-    if (currentSlide >= slides.length + slidesToScroll) setCurrentSlide(slidesToScroll);
-    if (currentSlide <= 0) setCurrentSlide(slides.length);
-  };
-
+  const [effectiveSlidesToScroll, setEffectiveSlidesToScroll] = useState(0);
   useLayoutEffect(() => {
-    setCurrentSlide(slidesToScroll);
-  }, [slidesToScroll]);
-
-  useLayoutEffect(() => {
-    if (!slideRef.current?.children[0]) return;
-
-    const observer = new ResizeObserver(([entry]) => {
-      const width = entry.target.getBoundingClientRect().width;
-      setSlideWidth(width);
-    });
-
-    observer.observe(slideRef.current.children[0]);
-    return () => observer.disconnect();
+    if (responsive.length > 0) {
+      const width = window.innerWidth;
+      const matchedSetting = responsive.sort((a, b) => b.breakpoint - a.breakpoint).find((r) => width >= r.breakpoint);
+      if (matchedSetting) {
+        setEffectiveSlidesToScroll(matchedSetting.slidesToScroll);
+      } else {
+        setEffectiveSlidesToScroll(slidesToScroll);
+      }
+    } else {
+      setEffectiveSlidesToScroll(slidesToScroll);
+    }
   }, []);
 
-  // auto: true일 때
-  useEffect(() => {
-    if (auto && !isTransitioning) {
-      timerRef.current = setTimeout(handleNextSlide, interval * 1000);
-    }
-    return () => {
-      if (timerRef.current) clearTimeout(timerRef.current);
-    };
-  }, [auto, isTransitioning, currentSlide]);
-
   return (
-    <>
-      <GlobalStyle />
-      <S.Container>
-        <S.ArrowWrapper className="left" $hasCustomArrow={!!leftArrow} onClick={handlePrevSlide}>
-          {leftArrow}
-        </S.ArrowWrapper>
-        <S.UlWrapper ref={containerRef}>
-          {auto && <S.ProgressBar key={currentSlide} $isTransitioning={isTransitioning} $interval={interval} />}
-          <S.SlideUl
-            ref={slideRef}
-            $currentSlide={currentSlide}
-            $isTransitioning={isTransitioning}
-            $slideWidth={slideWidth}
-            onTransitionEnd={handleTransitionEnd}
-          >
-            {DATA}
-          </S.SlideUl>
-        </S.UlWrapper>
-        <S.ArrowWrapper className="right" $hasCustomArrow={!!rightArrow} onClick={handleNextSlide}>
-          {rightArrow}
-        </S.ArrowWrapper>
-      </S.Container>
-    </>
+    <InnerInfiniteSlide
+      slidesToScroll={effectiveSlidesToScroll}
+      children={children}
+      leftArrow={leftArrow}
+      rightArrow={rightArrow}
+      auto={auto}
+      interval={interval}
+    />
   );
 };
