+++ src/lib/InnerInfiniteSlide/InnerInfiniteSlide.styles.tsx
