+++ src/lib/InnerInfiniteSlide/InnerInfiniteSlide.types.ts
@@ -1,14 +1,9 @@
-interface IResponsiveSetting {
-  breakpoint: number;
-  slidesToScroll: number;
-}
-
 export interface IProps {
-  slidesToScroll?: number;
+  slidesToShow?: number;
+  slidesToScroll: number;
   children: React.ReactNode;
   leftArrow?: React.ReactNode;
   rightArrow?: React.ReactNode;
   auto?: boolean;
   interval?: number; // auto일 때 시간 간격
-  responsive?: IResponsiveSetting[]; // 반응형
 }
